requirements.txt
django
pandas
django-bootstrap4
docxtpl
xlsxtpl
django-import-export
django-crispy-forms
django-admin-rangefilter
django-easy-select2
django-data-browser
django-grappelli
django-bootstrap-datepicker-plus